refactor: move timeout and delay to shared package
