+++ packages/shared/src/utils/index.ts
@@ -1,1 +1,3 @@
 export * from './parseURL'
+export * from './delay'
+export * from './timeout'
